feat: add workaround for ObjC modules used by `objc_library` targets  (#169)
- Add `src_type` to modules in the `swift_deps_index.json`. This value
identifies the type of source files in the module. This information is
used to resolve dependencies specifically as it relates to ObjC modules
consumed by `objc_library` targets.
- Implement `swift_objc_module_alias` macro. It defines a
`swift_library` that re-exports the specified modules defined in the
provided `deps`.
- For Swift targets implemented using ObjC, generate two Bazel targets.
One represents is the `objc_library`. The other is a
`swift_objc_module_alias` that is referenced by external repositories.
- Use `product_memberships` to determine whether the Bazel target(s) for
a Swift target should be publicly visible.

Related to #153.