feat: resolve Swift imports to Bazel targets via Swift package products (#197)
- Refactor `DependencyIndex` hiding the internals.
- Add `ResolveModulesToProducts` to `DependencyIndex`. This function
determines the best set of products to satisfy the modules required by a
`swift_xxx` declaration.
- Add `FindModules` to `DependencyIndex`. This function returns the list
of modules with a particular name. If given a slice of package
identities, it will filter the list to only include those modules in
that set.
- Add `package_identity` and `product_memberships` to the modules
written to the index.
- Add the list of identities for the direct dependencies for the
project. This is used to ensure that dependencies are not resolved to
indirect dependencies.
- Update the list of built-in Swift modules.
- Add custom package identity for modules defined in custom build files
(e.g. `http_archive` declarations).

Closes #194