fix: differentiate between GUI applications and command-line applications in Gazelle extension (#225)
Graphical user interface (GUI) applications can now use the `@main`
directive. The Gazlle extension attempts to identify the module type
based upon the presence of GUI-related imports.

Removed the UI test in the firebase abtesting example. It takes a long
time to run and does not contribute to the compilation check.

Related to #203.