fix: ignore phantom targets and ensure unique target label names (#401)
- The description manifests from SPM can contain undeclared targets
(i.e. phantom targets). We ignore these when constructing a Bazel build
file.
- It is critical that the Bazel label name for an SPM target not
conflict with the label name from a product. Typically, we do this by
using the path of the SPM target as the basis for the label name. This
PR fixes a bug where a non-unique label name was generated when an SPM
target's name and path were equal (i.e., no parent directory in the
path).

Related to #400.