feat: implement support for PascalCase module names with snake_case directory names (#534)
This PR introduces an optional Gazelle directive to convert module names
from snake_case to PascalCase for consistency with platform idioms.
Currently the module name is the same as the target name, derived
typically from the directory name.

Secondly, the current heuristic for detecting whether a directory is a
swift_test or swift_library / swift_binary checks for case sensitive,
"Test" or "Tests" directory names or "Test.swift" / "Tests.swift" file
names. This breaks if the directory name is in snake_case.

To address this issue I just made the checks case insensitive. I also
unified the DirSuffixes and fileSuffixes types into a PathSuffixes type
and added tests for both files and paths with both casings.