fix: Handle Xcode 15 swift package format (#631)
In Xcode 15 / Swift 5.9 Apple changed the output format of `swift
package dump-package` which rules_swift_package_manager parsed. This
caused the JSON deserialization to fail.

Specifically, the SourceControlLocation type changed from a single
RemoteLocation to an array of RemoteLocation(s) and these are now
objects with the key "urlString" instead of "url".

I added support for both the Xcode 14 / Swift 5.8 format as well as the
new Xcode 15 / Swift 5.9 format.
Whenever we drop support for Xcode 14, the support for the old format
can be removed.

I verified the change from my repo by using a local_path_override with
this patch and I verified that before swift_update_packages failed with
the error:
```
gazelle: index 0 expected to be type 'string', but was type 'map[string]interface {}'
```
While now it succeeds.