feat: Add directive to disable manual tag (#632)
This PR introduces a directive to disable the addition of the manual tag
to generated swift_library targets.

I understand the intent behind the tag and am ok with keeping it the
default. But for more advanced users who understand that they need to
pass the apple platform type configuration, it is useful to be able to
build globs of these targets without a dedicated build test target.