feat: add more built in frameworks (#750)
This resolves the first part of issue #742.

It adds support for finding built-in "frameworks" that are not included
as `.framework`s. In addition to the existing framework searching, we
now also search in the `usr/lib/swift` directory of the target SDK.

This allows including more builtin frameworks, such as `os` and
`Darwin`.

Support is added in a way where the frameworks and modules under
`/usr/lib/swift` are split. This is important for supporting
`obc_library` and `frameworks` attr. The swift modules are not valid
objective-c frameworks.

NOTE: this bumps the SDK version to 14.0 since I generated this file
using Xcode 15.0.0. The deletions seem reasonable as they have been
deprecated frameworks for a bit.

Depends on: #756