fix: ensure that the generated resource bundle finder class has a unique name (#803)
`rules_swift_package_manager` generates a resource bundle accessor for
ObjC code that has resources. The generated code includes a private
class that is used to find the resources. We need to ensure that the
finder class has a unique name.

- Generate a unique name for the ObjC resource bundle finder class.
- Update the `resources_example` to include a second Swift package that
has an ObjC resource module.
- Update the `README.md` for the `resources_example` with information
about one of the cases being tested.

Closes #798.