fix: filter out `maccatalyst` SPM platform (#829)
Bazel currently does not support `maccatalyst` as a platform. We were
mapping it to `ios` as they are Mac apps that use iOS frameworks.
However, this can lead to warnings if define values exist with the same
name mapped for `ios` and `maccatalyst` (e.g. Firebase).

This PR filters out `maccatalyst` conditions. 

Closes #801.