chore: upgrade bazel-gazelle to 0.35.0 in parent workspace (#828)
We, typically, try not to update dependencies listed in `MODULE.bazel`.
However, if the bazel-gazelle listed in `MODULE.bazel` gets out of sync
with the version of `bazel-gazelle` in `go.mod`, running `//:tidy` can
bork the build. So, this PR includes the changes from #826 and updates
`MODULE.bazel`. I also ran `//:tidy_all`.

Related to #826.

---------

Co-authored-by: Self-hosted Renovate Bot <361546+cgrindel-self-hosted-renovate[bot]@users.noreply.github.enterprise.com>