fix: copy Gazelle config to scope directives properly (#951)
A directive like `# gazelle:swift_generate_proto_libraries false` is
intended to apply to the current package and all subpackages. Currently,
though, it's being applied to all further packages that Gazelle parses.
This is because we have a shared reference to a single SwiftConfig
object that is being updated in place, so changes accidentally propagate
further than they should.

This PR instead makes a clone of the SwiftConfig for the current
package, based on the way it's done [in Gazelle
itself](https://github.com/bazelbuild/bazel-gazelle/blob/master/language/proto/config.go#L208-L210).