feat: no longer need to check in `swift_deps_index.json` file and other big updates (#957)
Lots of big changes in this PR:

- The SPM-related rules no longer depend upon a Swift deps index JSON
file.
- Remove the `from_file` module extension tag class, as the ruleset no
longer needs to read the Swift deps index file.
- Remove the `swift_update_packages` macro as it is no longer required.
- Remove support for `byName` references outside of a Swift package.
This was supported in early SPM versions. It was one of the driving
factors for having to pre-process the Swift dependencies and generate a
Swift deps index file. Removing this support reduces the complexity of
the `rules_swift_package_manager` implementation.
- Remove `update_repos` command from the Gazelle plugin.
- Add `swift_deps_info` repository rule to generate a Swift deps index
JSON file as part of the build, not in the Gazelle plugin. This index
file is used by the Gazelle plugin to resolve Swift module names to
their respective Bazel target.
- Introduce `from_package` module extension tag class telling the
ruleset where the `Package.swift` and `Package.resolved` files for the
workspace live. These are used to identify the Swift package transitive
dependencies for the project.
- Remove the `http_archive_ext_deps` example. The example existed to
demonstrate and test indexing of Swift targets that were downloaded
using `http_archive`. The Gazelle internals no longer provide some of
the information needed for this to work properly with bzlmod enabled.
Given that SPM has been widely adopted and there are some workarounds,
we are removing support for this functionality in the Gazelle plugin
when bzlmod is enabled.
- Add output to the example test runner that demarcates different phases
for the testing.

Related to #924.