fix: generate a `select` statement for each condition (#1173)
This is should fix #1111 . I had the same issue where the `select` in
`cops` would have multiple matching conditions, which I think isn't
something wrong. So in here I changed the code so it generates separate
selects for each option.

I'm not completely happy with the implementation and I still have to fix
the tests probably, but please tell me if you think I'm going in a
completely wrong direction

---------

Co-authored-by: Chuck Grindel <chuck.grindel@gmail.com>