fix: ignore unresolved packages in Package.swift
This fixes an issue where if a brand new dependency is added to the `Package.swift` when `Package.resolved` has not yet been updated, the Bazel build fails.
It fails because there is no `pin` available for the `source_control` dependency as the `Package.resolved` has not yet been updated.

Failing the build here causes issues for tools which are built & run with Bazel and operate on the same `Package.swift` they are built from.

For example, a command line tool built using `rules_swift_package_manager` which has a `Package.swift` for `swift-argument-parser`.
If a new dependency is added to the `Package.swift` but not yet used in any of the Bazel targets and a build is run the build will fail with:

```
File bazel-path/external/rules_swift_package_manager~/swiftpkg/bzlmod/swift_deps.bzl", line 152, column 25, in _declare_pkg_from_dependency
                commit = pin.state.revision,
Error: 'NoneType' value has no field or method 'state'
```

This change allows the build to progress by simply ignoring the unresolved dependencies until a `swift package resolve` is run by the user.