fix: set C/C++ language standards per target (#1217)
This PR refactors the creation of C, C++, ObjC, and ObjC++ targets. For
each C-like target in a Swift package, we now create a separate
`xxx_library` target based upon the type of the source file (e.g. C,
C++, ObjC, ObjC++, and assembly). These child libraries are added to a
parent library that represents the Swift package target.

The goal of this work was to support language standard flags in the
Bazel targets generated by `rules_swift_package_manager`. If a C
language standard is specified in the Swift package, it is applied to
the C and ObjC targets. If a C++ language standard is specified, it is
applied to the C++ and the ObjC++ targets.

> [!NOTE]
> The configuration for `rules_xcodeproj` in the Firebase example had to
be updated to use the sandbox. Without this change, we were seeing
duplicate definition errors. If you see these errors in your build, you
may need to do the same.

Closes #1079 .