chore: cleanup unresolved package handling
As I'm working on adding registry support I realized the registry is another case where it can be unresolved until a pin is present when resolved.
This PR cleans up the code added in #1223 to make it easier to ignore unresolved packages for both source control and registry dependencies in the future.