fix: upgrade `rules_bazel_integration_test` and remove `System` module from `//bzlmod:e2e_test` (#1517)
The test existed to test an edge case where someone creates a module
called `System`. The idea was that the `gazelle` plugin would pick the
custom one over the real `System` module. TBH, it is a bad idea to name
your module `System`. Something changed recently in the Swift toolchain
such that it no longer liked this test. So, we are removing it.

Closes #1514.