feat: Use GRPC from Swift Package Manager (#1532)
This is a re-do of
https://github.com/cgrindel/rules_swift_package_manager/pull/1019 which
was closed when @brentleyjones 's PR merged, rebased onto the main
branch.

This PR updates the GRPC examples to use a version of GRPC from RSPM.
This is important because without this change developers would get
compiler warnings or errors about duplicate symbols between the GRPC
version included with rules_swift and GRPC pulled in through a Swift
Package dependency.

I haven't been able to merge this PR previously because, before
@brentleyjones merged this fix:
https://github.com/cgrindel/rules_swift_package_manager/pull/1385

This is because the swift-atomics package (a dependency of grpc-swift)
crashes on ubuntu without this flag.

---------

Co-authored-by: Chuck Grindel <chuck.grindel@gmail.com>
Co-authored-by: Chuck Grindel <chuck@revealtech.ai>
Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>