feat: Support for SwiftSettings: swiftLanguageMode, enableUpcomingFeature, enableExperimentalFeature (#1423)
I'm trying to use some of the new GRPC and Protobuf package versions and
some of them have adopted upcoming and experimental features as well as
the Swift 6 language mode. These SwiftSettings are ignored in the
current master. This PR adds support for these features to the
repository rule.

## Test Plan

SwiftProtobuf uses the experimental StrictConcurrency feature and the
upcoming ExistentialAny feature, so it was a good test case for this
feature. This is the generated BUILD file with this branch in the
swift_protobuf example.

```python
logan@Logans-MacBook-Pro swift_protobuf_example % bazel query --output=build 
...
swift_library(
  name = "SwiftProtobuf.rspm",
  visibility = ["@swiftpkg_swift_protobuf//:__subpackages__"],
  tags = ["manual"],
  features = ["swift.experimental.StrictConcurrency=complete", "swift.upcoming.ExistentialAny"],
  data = ["@swiftpkg_swift_protobuf//:SwiftProtobuf.rspm_resource_bundle"],
  copts = ["-DSWIFT_PACKAGE"],
  module_name = "SwiftProtobuf",
  srcs = [...],
  always_include_developer_search_paths = True,
  alwayslink = True,
)
...
```

---------

Co-authored-by: Chuck Grindel <chuck@revealtech.ai>