fix: Always create cc_library targets if C/C++ files are present (#1594)
I ran into an issue were we couldn't build the
[InjectionNext](https://github.com/johnno1962/InjectionNext) package
because there was a C file that wasn't being included in the `srcs` list
of any generated target.

This was because most of the code is in Objective-C++, which creates a
target that doesn't include C source files.

This change always creates a dedicated `cc_library` target for any C/C++
code in the package.

---------

Co-authored-by: Chuck Grindel <chuck.grindel@gmail.com>
Co-authored-by: Chuck Grindel <chuck@revealtech.ai>
Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>