fix: `swift_package_tool` so that it works on Linux and MacOS (#1621)
The Renovate runs have been failing when trying to run
`//:update_swift_packages` on modified workspaces. The issue is that the
Swift worker processes behave differently on MacOS vs Linux. On MacOS,
the arguments are essentially passed through to `xcrun`. On Linux, it is
more of a barebones process runner.

Previously, we expected the worker to find `swift` automatically. This
does not work on Linux. This PR adds logic to find `swift` by attempting
to ask the worker. If that fails, fall back to looking in the `PATH`.

I also upgraded Swift to 6.1 on Ubuntu.

Closes #1591.

---------

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>