feat: Support `env` and `env_inherit` attrs in `swift_deps` module extension
Adds support for setting `env` explicitly or inherting them via `env_inherit`. This is required for some packages as some of them use `ProcessInfo.processInfo.environment` to alter behavior. This env is passed down to the relevant repository rules.

Also added a test case via `interesting_deps` where if this is not working the `Package.swift` provides invalid dependencies that should fail the tests. With this change the tests pass.

Signed-off-by: Luis Padron <heyluispadron@gmail.com>