fix: force symlink dereference when using file command (#1880)
On macOS, the `file` command dereferences symbolic links by default
(`--dereference` implied). On Linux it's the opposite
(`--no-dereference` implied).

We want the dereference behavior on both, otherwise `file` outputs
`symbolic link to Versions/Current/MyFrameworkBinary` on Linux when
checking the file type of an iOS framework (we then fallback to the
`otool` command which fails).